+++ src/utils/helpers.ts
@@ -1,0 +1,8 @@
+export const getParentDir = (filePath: string): string => {
+  const separator = filePath.includes('/') ? '/' : '\\';
+  const lastSeparatorIndex = filePath.lastIndexOf(separator);
+  if (lastSeparatorIndex === -1) {
+    return '';
+  }
+  return filePath.substring(0, lastSeparatorIndex);
+};
